v2.2.3 Release shutdownアクションの不具合修正 カスタムモニタリング動作中のダウンタイムの初期値を調整

+++ app/MainClass/MainForLauncher.php
@@ -202,9 +202,12 @@ finish:
             $log_writer = $this->log_writer;
             $log_writer($level, ['type' => $action, 'message' => $error_message, 'via' => 'CLI', 'who' => null, 'pid' => null]);
         }
-        if(file_exists($this->pid_path_for_launcher))
+        if(!isset($not_conflict) || (isset($not_conflict) && $not_conflict === true))
         {
-            unlink($this->pid_path_for_launcher);
+            if(file_exists($this->pid_path_for_launcher))
+            {
+                unlink($this->pid_path_for_launcher);
+            }
         }
         return;
     }
